changed xray

+++ Ripterms/Modules/AimAssist.cpp
@@ -19,12 +19,12 @@ void Ripterms::Modules::AimAssist::run()
 	thePlayer_position.y += cache->thePlayer.getEyeHeight();
 
 	Maths::Vector2d thePlayer_rotation = cache->thePlayer.getRotation();
-	float cropped_thePlayer_yaw = Ripterms::Maths::cropAngle180(thePlayer_rotation.x);
-	float cropped_thePlayer_pitch = thePlayer_rotation.y;
+	double cropped_thePlayer_yaw = Ripterms::Maths::cropAngle180(thePlayer_rotation.x);
+	double cropped_thePlayer_pitch = thePlayer_rotation.y;
 
 	EntityPlayer selected_target{};
-	float selected_target_YawToAdd = 0.0f;
-	float selected_target_PitchToAdd = 0.0f;
+	double selected_target_YawToAdd = 0.0;
+	double selected_target_PitchToAdd = 0.0;
 
 	for (EntityPlayer& target : Ripterms::cache->playerEntities.toVector<EntityPlayer>())
 	{
@@ -34,17 +34,17 @@ void Ripterms::Modules::AimAssist::run()
 		AxisAlignedBB targetBB = target.getBoundingBox();
 		Maths::Vector3d target_position
 		(
-			targetBB.getMinX() + (targetBB.getMaxX() - targetBB.getMinX()) / 2.0f,
-			targetBB.getMinY() + (targetBB.getMaxY() - targetBB.getMinY()) / 2.0f,
-			targetBB.getMinZ() + (targetBB.getMaxZ() - targetBB.getMinZ()) / 2.0f
+			targetBB.getMinX() + (targetBB.getMaxX() - targetBB.getMinX()) / 2.0,
+			targetBB.getMinY() + (targetBB.getMaxY() - targetBB.getMinY()) / 2.0,
+			targetBB.getMinZ() + (targetBB.getMaxZ() - targetBB.getMinZ()) / 2.0
 		);
 		target_position = target_position + target.getMovementVector(cache->timer.getRenderPartialTicks());
 
 		Maths::Vector2d target_required_rotation = Maths::getYawPitch(thePlayer_position, target_position);
-		float yawToAdd = target_required_rotation.x - cropped_thePlayer_yaw;
-		float pitchToAdd = target_required_rotation.y - cropped_thePlayer_pitch;
+		double yawToAdd = target_required_rotation.x - cropped_thePlayer_yaw;
+		double pitchToAdd = target_required_rotation.y - cropped_thePlayer_pitch;
 		yawToAdd = Ripterms::Maths::cropAngle180(yawToAdd);
-		float target_distance = (target_position - thePlayer_position).distance();
+		double target_distance = (target_position - thePlayer_position).distance();
 		if (target_distance <= max_distance && target_distance > 0.5 && std::abs(yawToAdd) <= max_angle)
 		{
 			if (prev_selected_target.isValid() && prev_selected_target.isEqualTo(target))
@@ -79,20 +79,20 @@ void Ripterms::Modules::AimAssist::run()
 			Ripterms::Maths::Vector3d(selected_target_bb.getMaxX(), selected_target_bb.getMinY(), selected_target_bb.getMaxZ()) + selected_target_mvmt_vec,
 			Ripterms::Maths::Vector3d(selected_target_bb.getMinX(), selected_target_bb.getMinY(), selected_target_bb.getMinZ()) + selected_target_mvmt_vec
 		};
-		float minYaw = 361.0f, maxYaw = -1.0f;
-		float minPitch = 0.0f, maxPitch = 0.0f;
+		double minYaw = 361.0, maxYaw = -1.0;
+		double minPitch = 0.0, maxPitch = 0.0;
 		for (const Ripterms::Maths::Vector3d& position : positionsToCheck)
 		{
 			Ripterms::Maths::Vector2d YawPitch = Maths::getYawPitch(thePlayer_position, position);
-			float Yaw = YawPitch.x;
-			if (Yaw < 0.0f) Yaw += 360.0f;
+			double Yaw = YawPitch.x;
+			if (Yaw < 0.0) Yaw += 360.0;
 			if (Yaw < minYaw) minYaw = Yaw;
 			if (Yaw > maxYaw) maxYaw = Yaw;
 			maxPitch = YawPitch.y;
 		}
 
-		if (minYaw > 180.0f) minYaw -= 360.0f;
-		if (maxYaw > 180.0f) maxYaw -= 360.0f;
+		if (minYaw > 180.0) minYaw -= 360.0;
+		if (maxYaw > 180.0) maxYaw -= 360.0;
 
 		Ripterms::Maths::Vector2d YawPitch = Maths::getYawPitch
 		(
@@ -102,63 +102,63 @@ void Ripterms::Modules::AimAssist::run()
 		minPitch = YawPitch.y;
 
 		//don't even ask, some random calculations that sometimes bypass anticheats
-		float maxDelta = std::abs(Ripterms::Maths::cropAngle180((maxYaw - minYaw))) / 2.0f;
-		float difference = std::abs(selected_target_YawToAdd) - maxDelta;
+		double maxDelta = std::abs(Ripterms::Maths::cropAngle180((maxYaw - minYaw))) / 2.0;
+		double difference = std::abs(selected_target_YawToAdd) - maxDelta;
 		if (std::abs(selected_target_YawToAdd) > maxDelta)
 		{
-			if (difference < 2.0f)
+			if (difference < 2.0)
 			{
-				thePlayer_rotation.x += difference * (selected_target_YawToAdd > 0.0f ? 1.0f : -1.0f);
+				thePlayer_rotation.x += difference * (selected_target_YawToAdd > 0.0 ? 1.0 : -1.0);
 			}
-			else if (difference >= 16.0f)
+			else if (difference >= 16.0)
 			{
-				std::uniform_int_distribution<> distr(int(400.0f * multiplier), int(800.0f * multiplier));
-				thePlayer_rotation.x += (float(distr(gen)) / 100.0f) * (selected_target_YawToAdd > 0.0f ? 1.0f : -1.0f);
+				std::uniform_int_distribution<> distr(int(400.0 * multiplier), int(800.0 * multiplier));
+				thePlayer_rotation.x += (double(distr(gen)) / 100.0) * (selected_target_YawToAdd > 0.0 ? 1.0 : -1.0);
 			}
-			else if (difference >= 8.0f)
+			else if (difference >= 8.0)
 			{
-				std::uniform_int_distribution<> distr(int(250.0f * multiplier), int(400.0f * multiplier));
-				thePlayer_rotation.x += (float(distr(gen)) / 100.0f) * (selected_target_YawToAdd > 0.0f ? 1.0f : -1.0f);
+				std::uniform_int_distribution<> distr(int(250.0 * multiplier), int(400.0 * multiplier));
+				thePlayer_rotation.x += (double(distr(gen)) / 100.0) * (selected_target_YawToAdd > 0.0 ? 1.0 : -1.0);
 			}
-			else if (difference >= 4.0f)
+			else if (difference >= 4.0)
 			{
-				std::uniform_int_distribution<> distr(int(100.0f * multiplier), int(200.0f * multiplier));
-				thePlayer_rotation.x += (float(distr(gen)) / 100.0f) * (selected_target_YawToAdd > 0.0f ? 1.0f : -1.0f);
+				std::uniform_int_distribution<> distr(int(100.0 * multiplier), int(200.0 * multiplier));
+				thePlayer_rotation.x += (double(distr(gen)) / 100.0) * (selected_target_YawToAdd > 0.0 ? 1.0 : -1.0);
 			}
 			else
 			{
-				std::uniform_int_distribution<> distr(int(25.0f * multiplier), int(100.0f * multiplier));
-				thePlayer_rotation.x += (float(distr(gen)) / 100.0f) * (selected_target_YawToAdd > 0.0f ? 1.0f : -1.0f);
+				std::uniform_int_distribution<> distr(int(25.0 * multiplier), int(100.0 * multiplier));
+				thePlayer_rotation.x += (double(distr(gen)) / 100.0) * (selected_target_YawToAdd > 0.0 ? 1.0 : -1.0);
 			}
 		}
 
-		float maxPitchDelta = std::abs(maxPitch - minPitch) / 2.0f;
+		double maxPitchDelta = std::abs(maxPitch - minPitch) / 2.0;
 		difference = std::abs(selected_target_PitchToAdd) - maxPitchDelta;
 		if (std::abs(selected_target_PitchToAdd) > maxPitchDelta)
 		{
-			if (difference < 2.0f)
+			if (difference < 2.0)
 			{
-				thePlayer_rotation.y += difference * (selected_target_PitchToAdd > 0.0f ? 1.0f : -1.0f);
+				thePlayer_rotation.y += difference * (selected_target_PitchToAdd > 0.0 ? 1.0 : -1.0);
 			}
-			else if (difference >= 16.0f)
+			else if (difference >= 16.0)
 			{
-				std::uniform_int_distribution<> distr(int(400.0f * multiplierPitch), int(800.0f * multiplierPitch));
-				thePlayer_rotation.y += (float(distr(gen)) / 100.0f) * (selected_target_PitchToAdd > 0.0f ? 1.0f : -1.0f);
+				std::uniform_int_distribution<> distr(int(400.0 * multiplierPitch), int(800.0 * multiplierPitch));
+				thePlayer_rotation.y += (double(distr(gen)) / 100.0) * (selected_target_PitchToAdd > 0.0 ? 1.0 : -1.0);
 			}
-			else if (difference >= 8.0f)
+			else if (difference >= 8.0)
 			{
-				std::uniform_int_distribution<> distr(int(250.0f * multiplierPitch), int(400.0f * multiplierPitch));
-				thePlayer_rotation.y += (float(distr(gen)) / 100.0f) * (selected_target_PitchToAdd > 0.0f ? 1.0f : -1.0f);
+				std::uniform_int_distribution<> distr(int(250.0 * multiplierPitch), int(400.0 * multiplierPitch));
+				thePlayer_rotation.y += (double(distr(gen)) / 100.0) * (selected_target_PitchToAdd > 0.0 ? 1.0 : -1.0);
 			}
-			else if (difference >= 4.0f)
+			else if (difference >= 4.0)
 			{
-				std::uniform_int_distribution<> distr(int(100.0f * multiplierPitch), int(200.0f * multiplierPitch));
-				thePlayer_rotation.y += (float(distr(gen)) / 100.0f) * (selected_target_PitchToAdd > 0.0f ? 1.0f : -1.0f);
+				std::uniform_int_distribution<> distr(int(100.0 * multiplierPitch), int(200.0 * multiplierPitch));
+				thePlayer_rotation.y += (double(distr(gen)) / 100.0) * (selected_target_PitchToAdd > 0.0 ? 1.0 : -1.0);
 			}
 			else
 			{
-				std::uniform_int_distribution<> distr(int(25.0f * multiplierPitch), int(100.0f * multiplierPitch));
-				thePlayer_rotation.y += (float(distr(gen)) / 100.0f) * (selected_target_PitchToAdd > 0.0f ? 1.0f : -1.0f);
+				std::uniform_int_distribution<> distr(int(25.0 * multiplierPitch), int(100.0 * multiplierPitch));
+				thePlayer_rotation.y += (double(distr(gen)) / 100.0) * (selected_target_PitchToAdd > 0.0 ? 1.0 : -1.0);
 			}
 		}
 		cache->thePlayer.setRotation(thePlayer_rotation);
